% Generated by build script
\name{hcgov_dataset}
\alias{hcgov_dataset}
\title{Get detailed info about a specific dataset}
\usage{
hcgov_dataset()
}
\arguments{
\item{id}{Dataset identifier (e.g. "e4rr-zk4i")}
}
\value{tibble: identifier, title, description, modified, access_level,}
\description{Get detailed info about a specific dataset}
