% Generated by build script
\name{hcgov_datasets}
\alias{hcgov_datasets}
\title{List available datasets on data.healthcare.gov}
\usage{
hcgov_datasets()
}
\arguments{
\item{page_size}{Number of results per page (default 20)}
\item{page}{Page number (default 1)}
}
\value{tibble: identifier, title, description, modified, access_level, keyword}
\description{List available datasets on data.healthcare.gov}
